Add Authentication for all Controler

+++ Trelo1/Controllers/BoardController.cs
@@ -6,6 +6,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Trelo1.Interfaces;
+using TreloBLL.ClaimsPrincipalExtensions;
 using TreloBLL.DtoModel;
 using TreloBLL.Interfaces;
 
@@ -53,8 +54,14 @@ namespace Trelo1.Controllers
         [Authorize(Roles = "Admin")]
         public async Task<IActionResult> AddUserToBoard(int boardId, SingleModel<int> userId)
         {
-            await _boardService.AddUserToBoard(userId.Value, boardId);
-            return Ok();
+            var currentUserId = User.GetUserId();
+            if (_appAuthentication.HasBoardAsses(currentUserId, boardId))
+            {
+                await _boardService.AddUserToBoard(userId.Value, boardId);
+                return Ok();
+            }
+
+            return StatusCode(401, "You haven't assess to this board");
         }
 
         [HttpDelete]
@@ -62,26 +69,36 @@ namespace Trelo1.Controllers
         [Authorize(Roles = "Admin")]
         public async Task<IActionResult> DeleteBoard(int boardId)
         {
-            bool hasDeleted = await _boardService.DeleteBoard(boardId);
-            if (hasDeleted)
+            var currentUserId = User.GetUserId();
+            if (_appAuthentication.HasBoardAsses(currentUserId, boardId))
             {
-                return Ok();
+                bool hasDeleted = await _boardService.DeleteBoard(boardId);
+                if (hasDeleted)
+                {
+                    return Ok();
+                }
+                return NoContent();
             }
-            return NoContent();
+            return StatusCode(401, "You haven't assess to this board");
+            
         }
         [HttpDelete]
         [Route("api/boards/{boardId}/users/")]
         [Authorize(Roles = "Admin")]
         public async Task<IActionResult> DeleteUserFromBoard(int boardId, SingleModel<int> userId)
         {
-            bool hasDeleted = await _boardService.DeleteUserFromBoard(userId.Value, boardId);
-            if (hasDeleted)
+            var currentUserId = User.GetUserId();
+            if (_appAuthentication.HasBoardAsses(currentUserId, boardId))
             {
-                return Ok();
-            }
+                bool hasDeleted = await _boardService.DeleteUserFromBoard(userId.Value, boardId);
+                if (hasDeleted)
+                {
+                    return Ok();
+                }
 
-            return NoContent();
-            
+                return NoContent();
+            }
+            return StatusCode(401, "You haven't assess to this board");
         }
     }
 }
