+++ Trelo1/Controllers/OrganizationController.cs
@@ -6,8 +6,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Trelo1.Interfaces;
+using TreloBLL.ClaimsPrincipalExtensions;
 using TreloBLL.DtoModel;
-
+using TreloBLL.Interfaces;
 
 namespace Trelo1.Controllers
 {
@@ -16,9 +17,11 @@ namespace Trelo1.Controllers
     public class OrganizationController : ControllerBase
     {
         private readonly IOrganizationService _organizationService;
-        public OrganizationController(IOrganizationService organizationService)
+        private readonly IAppAuthentication _appAuthentication;
+        public OrganizationController(IOrganizationService organizationService, IAppAuthentication appAuthentication)
         {
             _organizationService = organizationService;
+            _appAuthentication = appAuthentication;
         }
         [HttpPost]
         [Route("api/organizations/")]
@@ -31,15 +34,27 @@ namespace Trelo1.Controllers
         [Route("api/organizations/{orgId}")]
         public IActionResult DeleteOrg(int orgId)
         {
-            _organizationService.DeleteOrganization(orgId);
-            return Ok();
+            var curentUserId = User.GetUserId();
+            if (_appAuthentication.HasOrganizationAsses(curentUserId, orgId))
+            {
+                _organizationService.DeleteOrganization(orgId);
+                return Ok();
+            }
+
+            return StatusCode(401, "You haven't assess to this Org");
         }
         [HttpPost]
         [Route("api/organizations/{orgId}/add/boards/{boardId}/")]
         public IActionResult ChangeOganizationForBoard(int boardId, int orgId)
         {
-            _organizationService.AddBoardToOrg(boardId, orgId);
-            return Ok();
+            var curentUserId = User.GetUserId();
+            if (_appAuthentication.HasOrganizationAsses(curentUserId, orgId))
+            {
+                _organizationService.AddBoardToOrg(boardId, orgId); 
+                return Ok();
+            }
+
+            return StatusCode(401, "You haven't assess to this Org");
         }
     }
 }
